+++ LabelStoreMax/lib/app/networking/dio/interceptors/bearer_auth_interceptor.dart
@@ -16,7 +16,7 @@ class BearerAuthInterceptor extends Interceptor {
   }
 
   @override
-  void onError(DioError err, ErrorInterceptorHandler handler) {
+  void onError(DioException err, ErrorInterceptorHandler handler) {
     handler.next(err);
   }
 }