+++ LabelStoreMax/lib/config/providers.dart
@@ -1,3 +1,4 @@
+import '/app/providers/firebase_provider.dart';
 import 'package:flutter_app/app/providers/app_provider.dart';
 import 'package:flutter_app/app/providers/event_provider.dart';
 import 'package:flutter_app/app/providers/route_provider.dart';
@@ -17,4 +18,7 @@ final Map<Type, NyProvider> providers = {
   AppProvider: AppProvider(),
   RouteProvider: RouteProvider(),
   EventProvider: EventProvider(),
+  FirebaseProvider: FirebaseProvider(),
+
 };
+