+++ LabelStoreMax/lib/config/validation_rules.dart
@@ -8,9 +8,9 @@
 |--------------------------------------------------------------------------
 */
 
-final Map<Type, dynamic> validationRules = {
+final Map<String, dynamic> validationRules = {
   /// Example
-  // SimplePassword: (attribute) => SimplePassword(attribute)
+  // "simple_password": (attribute) => SimplePassword(attribute)
 };
 
 /// Example validation class